feat: make Extractors += extractor possible
...via redoing __iadd__(). Before that, my options were
Extractors.append(extractor) or Extractors += [extractor]. Less visual
noise is good :)

Note: Extractor itself is an Iterable (returns a list of k/v pair for
each field). Making a AppendableList(list[T]) was a struggle with the
type checker, as you can't use isinstance(x, T), so I was testing for
"not list" or "not Iterable" and this was prone to edge cases.